feat: implement server api  (#22)
This PR is iplementing a way to run server api without having a deamon process running at all time. It just boot the server and to the api call and shutdown after

This means we can now use parsed configs, user commands and agents without having to crawl ourselves the filesystem for them.

This should also address this PR #15

This should make the plugin more robust against opensode internal changes.