feat: use [server api's] instead or [opencode run] (#27)
This PR is refactoring the way this plugins interact with opencode.

It replaces the opencode run approach in favor of calling the api /session/{id}/message API on the server.

This allows to drop the xml templating for sending context in favor of sending message parts with the synthetic flag for sending context.

This also allows to reference files "natively" and reduce the "Read" tool usage in response. It also increase the speed of the response since the agent does not need to re-read the file.

This also makes the session compatible with the native opencode tui.