fix(streaming_renderer): lifecycle improvements
Trying to clean up the lifecycle management of streaming_renderer and
other objects by leverage the state listeners more.

Also move the improved autoscrolling code from output_renderer. Note,
there is an issue with permission replacement that leaves the cursor up
a few lines which stops scrolling so I'm still looking at that

Looked up the session compacted event and realized we don't need to
reset there but it would be nice to render something for the user.