fix(renderer): shift part lines
Lots going on here. First, I managed to get a session that delivered
multiple permission requests at the same time. I still not sure why that
happened.

However, that code exposed a bug in the part line shifting code. The
code was only shift parts that started after the new line end when it
needed to look for parts up to the old line end for the part being
replaced.