fix(core): restart opencode server on 3 cancels
If the user tries to cancel their request three times in a row, shutdown
the opencode server and then restart.

Getting the code working was a little bit tricky for two main issues:

1. If we try to do it on demand in api_client when the user sends a
prompt, then there's a race condition between that prompt getting
sent and when we get subscribed to server events. That race condition
can cause us to miss message/part events.

2. Some other listeners to opencode_server may trigger api_client calls
so we have to be extra careful to make sure base_url is set. That means
we can't rely solely on state.opencode_server notification to update
base_url but have to check it before using it.