refactor(event_manager): throttle event emission
Add new module, ThrottlingEmitter, that collects events and "drains"
them every drain_interval_ms. This allows us to batch events which lets
us process a group of them quickly. That, in turn, helps make sure we
don't overwhelm neovim (and particularly treesitter) with a ton of
events.

Before this change, the behavior we'd see is that the neovim ui could
get slow or hang for several/many seconds if we emitted too many events.
Disabling treesitter seemed to eliminate that behavior but it also
caused the buffers to lose their highlighting which wasn't a good user
experience.