fix(api): no restore_position on mention/slash
Restoring position was moving the cursor to the start of the input box
(the initial last saved position before we started insert mode) when
adding a mention rather than leaving the cursor where it is (i.e. where
the user is typing)

Since we're almost certainly already in insert mode, we don't want to
restore position when adding a mention / slash_command.