fix(files/snacks): restore win/mode/cursor on cancel
Snacks doesn't seem to restore the window/mode/cursor position when you
close the picker. We're already handling the case when you pick a file
but we weren't handling the case when you close the picker without
picking a file.

Now we save the window, mode, and cursor position and restore those if
we didn't pick a file