feat(rendering): incremental rendering (#62)
Introduces an incremental rendering that updates only the changed portion(s) of the buffer instead of re-rendering everything on every change. This reduces CPU usages for large sessions.

This PR also incorporate a functional test suite in `tests/manual` where you can record a run snapshot tests.

This is a major revision of the rendering system.


Co-authored-by:  Cameron Ring <cameron@cs.stanford.edu>