fix(renderer): parts occasionally not rendering
This was a subtle one and it only shows up when we're not collapsing
events so it wasn't caught by the unit tests. I've since modified the
unit tests to also run the replays without collapsing to catch bugs like
this in the future.

If the first version of a part didn't result in something being
rendered, future updates to that part would also not render anything.
This was particularly likely to happen with tools as the first version
of the part is often `step-start`. Because we didn't think this was a
new part, we called `_replace_part_in_buffer` but that exits quickly
because the part hadn't been rendered anywhere.

While we could try to call `_insert_part_to_buffer` in
`_replace_part_in_buffer` if the part hadn't been rendered, that would
miss some subtle cases around error handling so the correct fix is
consider the part new in `on_part_updated` if we haven't rendered it
before.