test(replay): fix full session replay
The issue is a bit subtle: the event notification about the session
changing from nil to an id gets wrapped in vim.schedule but that means
it gets delivered after we've loaded the fulls session, which causes us
to clear the whole session right after we just loaded it, leaving the
output blank.

It's possible that we might want to not wrap state notifications in
vim.schedule and expect listeners to handle cases where they can't be in
a fast context. The primary benefit of that would be guaranteed ordering
of events.

For example, in this case, the event that the session was set would've
been processed before we loaded the session so the issue wouldn't exist.